Conflicts:
	ChangeLog
	Makefile.in
	NEWS
	README
	aclocal.m4
	clamav-milter/Makefile.in
	clambc/Makefile.in
	clamconf/Makefile.in
	clamd/Makefile.in
	clamdscan/Makefile.in
	clamdscan/clamdscan.c
	clamdtop/Makefile.am
	clamdtop/Makefile.in
	clamscan/Makefile.in
	clamscan/manager.c
	clamsubmit/Makefile.in
	config/config.guess
	config/config.sub
	config/ltmain.sh
	config/missing
	configure
	configure.ac
	database/Makefile.in
	docs/Makefile.in
	docs/clamav-mirror-howto.pdf
	docs/clamdoc.pdf
	docs/clamdoc.tex
	docs/html/WARNINGS
	docs/html/clamdoc.html
	docs/html/footnode.html
	docs/html/img2.png
	docs/html/img3.png
	docs/html/img4.png
	docs/html/img5.png
	docs/html/img6.png
	docs/html/index.html
	docs/html/node1.html
	docs/html/node10.html
	docs/html/node11.html
	docs/html/node12.html
	docs/html/node13.html
	docs/html/node14.html
	docs/html/node15.html
	docs/html/node16.html
	docs/html/node17.html
	docs/html/node18.html
	docs/html/node19.html
	docs/html/node2.html
	docs/html/node20.html
	docs/html/node21.html
	docs/html/node22.html
	docs/html/node23.html
	docs/html/node24.html
	docs/html/node25.html
	docs/html/node26.html
	docs/html/node27.html
	docs/html/node28.html
	docs/html/node29.html
	docs/html/node3.html
	docs/html/node30.html
	docs/html/node31.html
	docs/html/node32.html
	docs/html/node33.html
	docs/html/node34.html
	docs/html/node35.html
	docs/html/node36.html
	docs/html/node37.html
	docs/html/node38.html
	docs/html/node39.html
	docs/html/node4.html
	docs/html/node40.html
	docs/html/node41.html
	docs/html/node42.html
	docs/html/node43.html
	docs/html/node44.html
	docs/html/node45.html
	docs/html/node46.html
	docs/html/node47.html
	docs/html/node48.html
	docs/html/node49.html
	docs/html/node5.html
	docs/html/node50.html
	docs/html/node51.html
	docs/html/node52.html
	docs/html/node53.html
	docs/html/node54.html
	docs/html/node55.html
	docs/html/node56.html
	docs/html/node57.html
	docs/html/node58.html
	docs/html/node59.html
	docs/html/node6.html
	docs/html/node60.html
	docs/html/node61.html
	docs/html/node62.html
	docs/html/node63.html
	docs/html/node64.html
	docs/html/node65.html
	docs/html/node66.html
	docs/html/node67.html
	docs/html/node68.html
	docs/html/node7.html
	docs/html/node8.html
	docs/html/node9.html
	docs/phishsigs_howto.pdf
	docs/signatures.pdf
	etc/Makefile.in
	etc/freshclam.conf.sample
	freshclam/Makefile.in
	freshclam/freshclam.c
	libclamav/Makefile.am
	libclamav/Makefile.in
	libclamav/c++/Makefile.in
	libclamav/c++/aclocal.m4
	libclamav/c++/bytecode2llvm.cpp
	libclamav/c++/config/config.guess
	libclamav/c++/config/config.sub
	libclamav/c++/config/ltmain.sh
	libclamav/c++/config/missing
	libclamav/crypto.h
	libclamav/others.c
	libclamav/www.c
	libltdl/Makefile.in
	m4/reorganization/version.m4
	shared/optparser.c
	sigtool/Makefile.in
	sigtool/sigtool.c
	test/Makefile.in
	unit_tests/Makefile.in
	unit_tests/check_clamav.c
	win32/Setup-x64/Setup-x64.vdproj
	win32/Setup-x86/Setup-x86.vdproj
	win32/res/common.rc

+++ libclamav/www.c
@@ -46,10 +46,6 @@
 
 #include "platform.h"
 
-#include <openssl/ssl.h>
-#include <openssl/err.h>
-#include "libclamav/crypto.h"
-
 #include "libclamav/others.h"
 #include "libclamav/clamav.h"
 #include "libclamav/www.h"
@@ -268,7 +264,7 @@ void submit_post(const char *host, const char *port, const char *method, const c
 
     cli_dbgmsg("stats - Connected to %s:%s\n", host, port);
 
-    if (send(sockfd, buf, strlen(buf), 0) != strlen(buf)) {
+    if ((size_t)send(sockfd, buf, strlen(buf), 0) != (size_t)strlen(buf)) {
         closesocket(sockfd);
         free(buf);
         return;
